requirements.txt
pyspark>=3.5.0
pandas
seaborn
matplotlib
folium
pyyaml
requests
tqdm
streamlit        
pyarrow          
fastparquet     

streamlit-folium
geopandas

pytest